Refactor 1
